testing: hoist shared R<>W/K<>P helpers into common, drop dead code
Both bridge pairs carried byte/near-identical copies of the network-spawn and
genesis-override helpers. Move them into common/utils.rs:

  - spawn_with_retry, now generic over the provider spawn fn (docker
    get_spawn_fn() vs native Provider::Native.get_spawn_fn());
  - bridge_hub_balances (was bridge_hub_balances_override), with the per-pair
    endowment and sovereign funding as parameters;
  - global_settings, config_errs and wait_for_native_increase.

Remove three helpers with no callers: run_relayer_to_completion,
wait_for_block_height and parse_account.
