+++ testing/zombienet-sdk-tests/tests/common/utils.rs
@@ -1,15 +1,17 @@
 // Copyright (C) Parity Technologies (UK) Ltd.
 // SPDX-License-Identifier: Apache-2.0
 
-//! Generic, runtime-agnostic `subxt` helpers shared by every bridge pair: reorg-tolerant extrinsic
-//! submission, balance/header queries, block-height waits and a polling retry combinator.
+//! Generic, runtime-agnostic helpers shared by every bridge pair: reorg-tolerant extrinsic
+//! submission, balance/header queries, a polling retry combinator, and the zombienet
+//! network-spawn / genesis-override helpers their `environment` modules build on.
 
 use anyhow::anyhow;
 use codec::Decode;
 use std::{future::Future, time::Duration};
 use subxt::{config::DefaultExtrinsicParamsBuilder, tx::Payload, OnlineClient, PolkadotConfig};
-use subxt_signer::sr25519::Keypair;
+use subxt_signer::sr25519::{dev, Keypair};
 use tokio::time::{sleep, timeout_at, Instant};
+use zombienet_sdk::{GlobalSettingsBuilder, LocalFileSystem, Network, NetworkConfig};
 
 /// Whether `e` is a transient reorg failure on the reorgy asset/bridge hubs, safe to retry (nothing
 /// durable was committed). subxt only exposes these as `Display` text, so we match on it: a
@@ -199,28 +201,6 @@ pub async fn best_finalized_bridged_header(
 	Ok(decoded.map(|(number, _hash)| number))
 }
 
-/// Waits until `client` reports a best block of at least `height`, or `timeout` elapses.
-///
-/// Uses best (not finalized) blocks: confirms block production without depending on the finality
-/// chain.
-// Currently unused: `init_bridge` gates on the bridge hubs' first *finalized* block instead (a
-// stronger, faster signal). Kept as a helper for ad-hoc best-block waits.
-#[allow(dead_code)]
-pub async fn wait_for_block_height(
-	client: &OnlineClient<PolkadotConfig>,
-	height: u32,
-	timeout: Duration,
-) -> Result<(), anyhow::Error> {
-	let mut sub = client.blocks().subscribe_best().await?;
-	let deadline = Instant::now() + timeout;
-	while let Ok(Some(block)) = timeout_at(deadline, sub.next()).await {
-		if block?.number() >= height {
-			return Ok(());
-		}
-	}
-	Err(anyhow!("timeout waiting for block height {height}"))
-}
-
 /// Waits until `client` reports a *finalized* block of at least `height`, or `timeout` elapses.
 ///
 /// Gates `init-bridge`: a freshly-started bridge hub reorgs heavily at the tip, so init waits for
@@ -298,8 +278,103 @@ pub fn dev_public(keypair: &Keypair) -> [u8; 32] {
 	keypair.public_key().0
 }
 
-#[allow(dead_code)]
-fn parse_account(ss58: &str) -> Result<subxt::utils::AccountId32, anyhow::Error> {
-	ss58.parse::<subxt::utils::AccountId32>()
-		.map_err(|e| anyhow!("invalid SS58 account {ss58}: {e:?}"))
+// ---------------------------------------------------------------------------------------------
+// Network spawn & genesis helpers (shared by every bridge pair's `environment`).
+// ---------------------------------------------------------------------------------------------
+
+/// Shared zombienet global settings: disable `tear_down_on_failure` so a transient, load-induced
+/// node-monitor timeout doesn't tear the network down, and allow 600s per node spawn.
+pub fn global_settings(settings: GlobalSettingsBuilder) -> GlobalSettingsBuilder {
+	settings.with_tear_down_on_failure(false).with_node_spawn_timeout(600)
+}
+
+/// Collapses zombienet network-config build errors into a single `anyhow::Error`.
+pub fn config_errs(errs: Vec<anyhow::Error>) -> anyhow::Error {
+	anyhow!(
+		"network config errors: {}",
+		errs.iter().map(|e| e.to_string()).collect::<Vec<_>>().join("; ")
+	)
+}
+
+/// Spawns one network with the given `spawn_fn` (native or docker provider), retrying the flaky
+/// zombienet chain-spec panic (`is_raw()` unwraps a truncated spec read and panics with `EOF while
+/// parsing ...`). Rebuilds the config each attempt for a fresh namespace; a genuine `Err` is
+/// retried too and surfaces after the last attempt.
+pub async fn spawn_with_retry<F, Fut, E>(
+	spawn_fn: F,
+	config_fn: impl Fn() -> Result<NetworkConfig, anyhow::Error>,
+	name: &str,
+) -> Result<Network<LocalFileSystem>, anyhow::Error>
+where
+	F: Fn(NetworkConfig) -> Fut,
+	Fut: Future<Output = Result<Network<LocalFileSystem>, E>> + Send + 'static,
+	E: std::fmt::Display + Send + 'static,
+{
+	const MAX_ATTEMPTS: usize = 3;
+	let mut last_err = String::new();
+	for attempt in 1..=MAX_ATTEMPTS {
+		let config = config_fn()?;
+		// Run in a task so a panic in zombienet is captured as a `JoinError`, not unwound.
+		match tokio::spawn(spawn_fn(config)).await {
+			Ok(Ok(network)) => return Ok(network),
+			Ok(Err(e)) => {
+				last_err = e.to_string();
+				log::warn!(
+					"{name} network spawn attempt {attempt}/{MAX_ATTEMPTS} failed: {last_err}"
+				);
+			},
+			Err(join_err) if join_err.is_panic() => {
+				let panic = join_err.into_panic();
+				last_err = panic
+					.downcast_ref::<&str>()
+					.map(|s| s.to_string())
+					.or_else(|| panic.downcast_ref::<String>().cloned())
+					.unwrap_or_else(|| "unknown panic".to_string());
+				log::warn!(
+					"{name} network spawn attempt {attempt}/{MAX_ATTEMPTS} panicked inside \
+					 zombienet (likely the chain-spec truncation flake); retrying: {last_err}"
+				);
+			},
+			Err(join_err) => return Err(anyhow!("{name} network spawn task cancelled: {join_err}")),
+		}
+	}
+	Err(anyhow!("{name} network spawn failed after {MAX_ATTEMPTS} attempts: {last_err}"))
+}
+
+/// Genesis `balances` override for a bridge hub. `with_genesis_overrides` *replaces* the
+/// `balances.balances` array (it doesn't append), so we re-list the well-known dev accounts
+/// (collators + relayer signers) with `endowment`, then fund the `sovereign_accounts` (sovereign /
+/// reward accounts) with `sovereign_funding`.
+pub fn bridge_hub_balances(
+	endowment: u128,
+	sovereign_accounts: &[&str],
+	sovereign_funding: u128,
+) -> serde_json::Value {
+	let mut balances: Vec<serde_json::Value> =
+		[dev::alice(), dev::bob(), dev::charlie(), dev::dave(), dev::eve(), dev::ferdie()]
+			.iter()
+			.map(|k| serde_json::json!([dev_account(k).to_string(), endowment]))
+			.collect();
+	for account in sovereign_accounts {
+		balances.push(serde_json::json!([*account, sovereign_funding]));
+	}
+	serde_json::json!({ "balances": { "balances": balances } })
+}
+
+/// Waits (up to 10 minutes) until the native free balance of `account` on `client` exceeds
+/// `initial + min_increase`.
+pub async fn wait_for_native_increase(
+	client: &OnlineClient<PolkadotConfig>,
+	account: [u8; 32],
+	initial: u128,
+	min_increase: u128,
+) -> Result<(), anyhow::Error> {
+	retry_until(Duration::from_secs(600), || {
+		let client = client.clone();
+		async move {
+			let balance = free_balance_at(&client, account).await?;
+			Ok((balance > initial + min_increase).then_some(()))
+		}
+	})
+	.await
 }
