+++ testing/zombienet-sdk-tests/tests/rococo_westend/asset_transfer.rs
@@ -11,13 +11,15 @@
 //!   * the finality/parachain relayers (`//Charlie` / `//Dave`) keep a constant balance, because
 //!     their transactions are free.
 
-use crate::rococo_westend::{
-	assert_relayer_balances_unchanged, asset_hub_rococo, asset_hub_westend,
-	bridge_hub_rococo_relayer_reward, bridge_hub_westend_relayer_reward, dev_account, dev_public,
-	retry_until, BridgeTestEnv,
+use crate::{
+	common::utils::wait_for_native_increase,
+	rococo_westend::{
+		assert_relayer_balances_unchanged, asset_hub_rococo, asset_hub_westend,
+		bridge_hub_rococo_relayer_reward, bridge_hub_westend_relayer_reward, dev_account,
+		dev_public, retry_until, BridgeTestEnv,
+	},
 };
 use std::time::Duration;
-use subxt::{OnlineClient, PolkadotConfig};
 use subxt_signer::sr25519::dev;
 
 const FIVE_UNITS: u128 = 5_000_000_000_000;
@@ -160,22 +162,4 @@ async fn asset_transfer_works() -> Result<(), anyhow::Error> {
 	assert_relayer_balances_unchanged(&bhr, &bhw, charlie, dave).await?;
 
 	Ok(())
-}
-
-/// Waits (up to 10 minutes) until the native free balance of `account` on `client` exceeds
-/// `initial + min_increase`.
-async fn wait_for_native_increase(
-	client: &OnlineClient<PolkadotConfig>,
-	account: [u8; 32],
-	initial: u128,
-	min_increase: u128,
-) -> Result<(), anyhow::Error> {
-	retry_until(Duration::from_secs(600), || {
-		let client = client.clone();
-		async move {
-			let balance = crate::rococo_westend::free_balance_at(&client, account).await?;
-			Ok((balance > initial + min_increase).then_some(()))
-		}
-	})
-	.await
 }
